smb 2

+++ init.sh.08.samba.sh
@@ -1,7 +1,10 @@
 yum install -yt samba
 yum install -yt samba-client
 
-mv /etc/samba/smb.conf /etc/samba/smb.conf.bkp
+if [[ -f "/etc/samba/smb.conf" ]]; then
+	mv /etc/samba/smb.conf /etc/samba/smb.conf.bkp
+fi
+
 cp ~/ec2init/mods/smb.conf.new /etc/samba/smb.conf
 
 service smb start
